(refactor) Centralize FHIR patient mapping logic in search app
Extracts duplicated FHIR patient mapping logic into a centralized utility
function in `fhir-mapper.ts`. Improves FHIR compliance by:

- Using standard `fhir.Patient` type
- Properly structuring telecom data
- Adding required `resourceType` field
- Better handling of null/undefined values

This change reduces code duplication across patient banner components and makes future
updates to mapping logic easier to maintain.
