+++ packages/esm-patient-search-app/src/patient-search-page/patient-banner/banner/patient-banner.component.tsx
@@ -1,6 +1,5 @@
 import React, { useCallback, useState, useMemo } from 'react';
 import classNames from 'classnames';
-import { v4 as uuidv4 } from 'uuid';
 import { SkeletonIcon, SkeletonText } from '@carbon/react';
 import {
   ConfigurableLink,
@@ -15,8 +14,9 @@ import {
   useVisit,
 } from '@openmrs/esm-framework';
 import { type PatientSearchConfig } from '../../../config-schema';
-import { type FHIRPatientType, type SearchedPatient } from '../../../types';
+import { type SearchedPatient } from '../../../types';
 import { usePatientSearchContext } from '../../../patient-search-context';
+import { mapToFhirPatient } from '../../../utils/fhir-mapper';
 import styles from './patient-banner.scss';
 
 interface ClickablePatientContainerProps {
@@ -61,53 +61,7 @@ const PatientBanner: React.FC<PatientBannerProps> = ({ patient, patientUuid, hid
     setShowContactDetails((value) => !value);
   }, []);
 
-  const fhirMappedPatient: FHIRPatientType = useMemo(() => {
-    const preferredAddress = patient.person.addresses?.find((address) => address.preferred);
-    const addressId = uuidv4();
-    const nameId = uuidv4();
-
-    return {
-      address: preferredAddress
-        ? [
-            {
-              id: addressId,
-              city: preferredAddress.cityVillage,
-              country: preferredAddress.country,
-              postalCode: preferredAddress.postalCode,
-              state: preferredAddress.stateProvince,
-              use: 'home',
-            },
-          ]
-        : [],
-      birthDate: patient.person.birthdate,
-      deceasedBoolean: patient.person.dead,
-      deceasedDateTime: patient.person.deathDate,
-      gender: getGender(patient.person.gender),
-      id: patient.uuid,
-      identifier: patient.identifiers.map((identifier) => ({
-        id: identifier.uuid,
-        type: {
-          coding: [
-            {
-              code: identifier.identifierType.uuid,
-            },
-          ],
-          text: identifier.identifierType.display,
-        },
-        use: 'official',
-        value: identifier.identifier,
-      })),
-      name: [
-        {
-          family: patient.person.personName.familyName,
-          given: [patient.person.personName.givenName, patient.person.personName.middleName],
-          id: nameId,
-          text: patient.person.personName.display,
-        },
-      ],
-      telecom: patient.attributes?.filter((attribute) => attribute.attributeType.display === 'Telephone Number'),
-    };
-  }, [patient]);
+  const fhirMappedPatient: fhir.Patient = useMemo(() => mapToFhirPatient(patient), [patient]);
 
   return (
     <>
